requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0

